adding conditional rendering for different headers
